Adding Gulp tasks for CSS,HTML and JS
