+++ test-ui/src/main/resources/static-dev/app/user/directives/user.directives.js
@@ -1,0 +1,5 @@
+(function (module) {
+
+    'use strict';
+
+})(angular.module('user'));
