+++ test-ui/src/main/resources/static-dev/app/user/user.js
@@ -1,0 +1,5 @@
+(function () {
+
+    'use strict';
+
+})(angular.module('user', []));
