+++ test-ui/src/main/resources/static-dev/app/user/views/list.html
@@ -1,0 +1,52 @@
+<div class="row divpaddingTop">
+    <div class="panel panel-default">
+        <div class="panel-heading">
+            <h3 class="panel-title">Usuários</h3>
+        </div>
+        <div class="panel-body">
+
+
+            <div class="mt-12 col-md-12">
+                <div class="row">
+                    <a href="#user/add">
+                        <button type="button" class="btn btn-default btn-success pull-right">
+                            Cadastrar usuário
+                        </button>
+                    </a>
+                </div>
+            </div>
+
+            <div class="row">
+                <table class="table table-striped table-bordered box list-table">
+                    <thead>
+                    <th>Nome</th>
+                    <th>E-mail</th>
+                    <th>Telefone</th>
+                    <th>Sexo</th>
+                    <th>Ações</th>
+                    </thead>
+                    <tbody>
+                    <tr ng-repeat="user in users">
+                        <td>{{user.name}}</td>
+                        <td>{{user.email}}</td>
+                        <td>{{user.phone}}</td>
+                        <td>{{user.sex}}</td>
+                        <td>
+                            <a ui-sref="user/edit({userId:user.id})">
+                                <button type="button" class="btn btn-default btn-xs btn-primary">
+                                   Alterar
+                                </button>
+                            </a>
+
+                            <button type="button" class="btn btn-default btn-xs btn-danger confirm"  id="{{$index}}" ng-click="removeUser(user.id);">
+                                Remover
+                            </button>
+                        </td>
+                    </tr>
+                    </tbody>
+                </table>
+            </div>
+        </div>
+
+    </div>
+</div>
