+++ test-ui/src/main/resources/static-dev/index.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="pt-BR" ng-app="test">
+
+<head>
+
+<meta name="viewport"
+	content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+<meta charset="UTF-8">
+
+<title>Test Full Stack</title>
+
+<link rel="stylesheet" href="/vendors/bootstrap/bootstrap.min.css">
+<link rel="stylesheet" type="text/css" href="/css/app.css" />
+
+</head>
+
+<body ng-app="test">
+
+	<div class="container">
+		<div class="row">
+			<div>
+				<div id="mainContext" ui-view></div>
+			</div>
+		</div>
+	</div>
+
+	<script src="/vendors/jquery/jquery-3.1.1.min.js"></script>
+	<script src="/vendors/bootstrap/bootstrap.min.js"></script>
+	<script src="/vendors/angular/angular.min.js"></script>
+	<script src="/vendors/angular/angular-ui-router.min.js"></script>
+
+	<script src="/app/app.js"></script>
+
+	<!-- module user -->
+	<script src="/app/user/user.js"></script>
+	<script src="/app/user/directives/user.directives.js"></script>
+	<script src="/app/user/interceptors/user-interceptor.js"></script>
+	<script src="/app/user/services/user-service.js"></script>
+	<script src="/app/user/controllers/user-controller.js"></script>
+
+
+</body>
+
+</html>
