+++ test-ui/src/main/resources/static-dev/package.json
@@ -18,8 +18,11 @@
   "homepage": "https://github.com/omotor/test-fullstack#readme",
   "devDependencies": {
     "gulp": "^3.9.1",
+    "gulp-htmlmin": "^3.0.0",
     "gulp-less": "^3.3.0",
+    "gulp-minify-css": "^1.2.4",
     "gulp-uglify": "^2.1.2",
+    "gulp-util": "^3.0.8",
     "require-dir": "^0.3.1"
   }
 }
