+++ test-ui/src/main/resources/static/app/app.js
@@ -1,29 +1,1 @@
-(function (module) {
-
-	'use strict';
-
-    module.config(['$urlRouterProvider', '$stateProvider',
-		function ($urlRouterProvider, $stateProvider) {
-
-			$urlRouterProvider.otherwise('/user');
-
-			$stateProvider.state('user', {
-				url: '/user',
-				templateUrl : 'app/user/views/list.html',
-				controller : 'UserController'
-			}).state('user/add', {
-				url: '/user/add',
-				templateUrl : 'app/user/views/user.html',
-				controller : 'UserController'
-			}).state('user/edit', {
-                url: '/user/:userId',
-				templateUrl : 'app/user/views/user.html',
-                controller : 'UserController'
-            });
-}]);
-
-    module.run(['$rootScope', '$state',
-		function ($rootScope, $state) {}]);
-
-})(angular.module('test', ['ui.router', 'user']));
-
+!function(e){"use strict";e.config(["$urlRouterProvider","$stateProvider",function(e,r){e.otherwise("/user"),r.state("user",{url:"/user",templateUrl:"app/user/views/list.html",controller:"UserController"}).state("user/add",{url:"/user/add",templateUrl:"app/user/views/user.html",controller:"UserController"}).state("user/edit",{url:"/user/:userId",templateUrl:"app/user/views/user.html",controller:"UserController"})}]),e.run(["$rootScope","$state",function(e,r){}])}(angular.module("test",["ui.router","user"]));
