--- test-ui/src/main/resources/static/app/consts.js
@@ -1,2 +1,0 @@
-var ENDPOINT="api/";
-var API_USER="user"
