+++ test-ui/src/main/resources/static/app/user/controllers/user-controller.js
@@ -1,59 +1,1 @@
-(function (module) {
-
-    'use strict';
-
-    module.controller('UserController', ['$scope', '$state', 'UserService', function ($scope, $state,  userService) {
-
-         userService.loadUsers().then(function(users) {
-            $scope.users = users;
-        });
-
-         $scope.$state = $state;
-
-         $scope.removeUser = function(id) {
-
-             userService.removeUser(id).then(function() {
-                 userService.loadUsers().then(function(users) {
-                     $scope.users = users;
-                 });
-             });
-
-         };
-
-         $scope.saveUser = function(form) {
-
-             $scope.submitted = true;
-
-             if (form.$invalid) {
-                 return;
-             }
-
-             console.log(form);
-
-             var user = {
-                 id: form.id ? form.id.$modelValue : null,
-                 name: form.name.$modelValue,
-                 email: form.email.$modelValue,
-                 phone: form.phone.$modelValue,
-                 sex: form.sex.$modelValue
-             };
-
-             userService.saveUser(user).then(function() {
-                 $state.go('user');
-             });
-
-         };
-
-         $scope.loadUser = function() {
-             if ($state.params.userId) {
-                 userService.listUser($state.params.userId).then(function (user) {
-                     $scope.user = user;
-                 });
-             }
-         };
-
-         $scope.loadUser();
-    }]);
-
-
-})(angular.module('user'));
+!function(e){"use strict";e.controller("UserController",["$scope","$state","UserService",function(e,s,n){n.loadUsers().then(function(s){e.users=s}),e.$state=s,e.removeUser=function(s){n.removeUser(s).then(function(){n.loadUsers().then(function(s){e.users=s})})},e.saveUser=function(o){if(e.submitted=!0,!o.$invalid){console.log(o);var r={id:o.id?o.id.$modelValue:null,name:o.name.$modelValue,email:o.email.$modelValue,phone:o.phone.$modelValue,sex:o.sex.$modelValue};n.saveUser(r).then(function(){s.go("user")})}},e.loadUser=function(){s.params.userId&&n.listUser(s.params.userId).then(function(s){e.user=s})},e.loadUser()}])}(angular.module("user"));
