+++ test-ui/src/main/resources/static/app/user/directives/user.directives.js
@@ -1,5 +1,1 @@
-(function (module) {
-
-    'use strict';
-
-})(angular.module('user'));
+!function(u){"use strict"}(angular.module("user"));
