+++ test-ui/src/main/resources/static/app/user/services/user-service.js
@@ -1,103 +1,1 @@
-(function (module) {
-
-    'use strict';
-
-    module.service('UserService', ['$http', function ($http) {
-
-        function handleError(response) {
-
-        }
-
-        function handleSuccess(response) {
-            return (response.data);
-        }
-
-        var saveUser = function (user) {
-
-            var request = "";
-
-            if (user.id) {
-
-                request = $http({
-                    method: "PUT",
-                    url: ENDPOINT + API_USER + "/" + user.id,
-                    headers: {
-                        'Accept': 'application/json',
-                        'Content-Type': 'application/json'
-                    },
-                    data: JSON.stringify(user)
-                });
-
-            } else {
-
-                request = $http({
-                    method: "POST",
-                    url: ENDPOINT + API_USER,
-                    headers: {
-                        'Accept': 'application/json',
-                        'Content-Type': 'application/json'
-                    },
-                    data: JSON.stringify(user)
-                });
-
-            }
-
-            return (request.then(handleSuccess, handleError));
-        };
-
-        var listUser = function (id) {
-
-            var request = $http({
-                method: "GET",
-                url: ENDPOINT + API_USER + "/" + id,
-                headers: {
-                    'Accept': 'application/json',
-                    'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8'
-                }
-            });
-
-            return (request.then(handleSuccess, handleError));
-
-        };
-
-        var loadUsers = function (user) {
-
-            var request = $http({
-                method: "GET",
-                url: ENDPOINT + API_USER,
-                headers: {
-                    'Accept': 'application/json'
-                }
-            });
-
-            return (request.then(handleSuccess, handleError));
-
-        };
-
-        var removeUser = function (id) {
-
-            var request = $http({
-                method: "DELETE",
-                url: ENDPOINT + API_USER + "/" + id,
-                headers: {
-                    'Accept': 'application/json',
-                    'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8'
-                }
-            });
-
-            return (request.then(handleSuccess, handleError));
-
-        };
-
-        return {
-            loadUsers: loadUsers,
-            listUser: listUser,
-            saveUser: saveUser,
-            removeUser: removeUser
-        };
-
-    }])
-
-})(angular.module('user'));
-
-
+!function(e){"use strict";e.service("UserService",["$http",function(e){function t(e){}function n(e){return e.data}var r=function(r){var a="";return a=e(r.id?{method:"PUT",url:"api/user/"+r.id,headers:{Accept:"application/json","Content-Type":"application/json"},data:JSON.stringify(r)}:{method:"POST",url:"api/user",headers:{Accept:"application/json","Content-Type":"application/json"},data:JSON.stringify(r)}),a.then(n,t)},a=function(r){return e({method:"GET",url:"api/user/"+r,headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded; charset=UTF-8"}}).then(n,t)};return{loadUsers:function(r){return e({method:"GET",url:"api/user",headers:{Accept:"application/json"}}).then(n,t)},listUser:a,saveUser:r,removeUser:function(r){return e({method:"DELETE",url:"api/user/"+r,headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded; charset=UTF-8"}}).then(n,t)}}}])}(angular.module("user"));
