+++ test-ui/src/main/resources/static/app/user/user.js
@@ -1,5 +1,1 @@
-(function () {
-
-    'use strict';
-
-})(angular.module('user', []));
+!function(){"use strict"}(angular.module("user",[]));
