+++ test-ui/src/main/resources/static/app/user/views/user.html
@@ -1,67 +1,1 @@
-<div class="divpaddingTop"></div>
-
-<form name="userForm" novalidate>
-
-    <div class="panel panel-default">
-
-        <div class="panel-heading">
-            <h3 class="panel-title">Cadastro de Usuário</h3>
-        </div>
-
-        <div class="panel-body">
-
-
-            <div class="panel panel-warning">
-                <div class="panel-heading">
-                    <h3 class="panel-title">Dados pessoais</h3>
-                </div>
-                <div class="panel-body">
-
-                    <input type="hidden" name="id" ng-model="user.id" />
-
-                    <div class="form-group">
-                        <label name="name" class="nome required control-label">Nome</label>
-                        <input type="text" name="name" ng-model="user.name" class="string required form-control" placeholder="Nome completo" maxlength="300" required />
-                        <p ng-show="userForm.name.$invalid && !userForm.name.$pristine" class="help-block">O nome é obrigatório.</p>
-                    </div>
-
-
-                    <div class="form-group">
-                        <label name="email" class="email required control-label">Email</label>
-                        <input type="text" name="email" ng-model="user.email" class="string email required form-control" placeholder="correio eletrônico/email" maxlength="300" required/>
-                        <p ng-show="userForm.email.$invalid && !userForm.email.$pristine" class="help-block">O e-mail é obrigatório.</p>
-                    </div>
-
-                    <div class="form-group">
-                        <label name="phone" class="email required control-label">Telefone</label>
-                        <input type="text" name="phone" ng-model="user.phone" class="string required form-control" placeholder="Telefone" maxlength="255" required/>
-                        <p ng-show="userForm.phone.$invalid && !userForm.phone.$pristine" class="help-block">O telefone é obrigatório.</p>
-                    </div>
-
-                    <div class="form-group">
-                        <label name="sex" class="email required control-label">Sexo</label>
-                        <select name="sex" ng-model="user.sex" class="string required form-control" required>
-                            <option value=""> - </option>
-                            <option value="Masculino">Masculino</option>
-                            <option value="Feminino">Feminino</option>
-                        </select>
-
-                        <p ng-show="userForm.sex.$invalid && !userForm.sex.$pristine" class="help-block">O sexo é obrigatório.</p>
-                    </div>
-
-                </div>
-
-            </div>
-        </div>
-    </div>
-
-    <div class="footer">
-        <div class="row">
-            <a href="#home">
-                <button class="btn btn-default btn-primary pull-left">Voltar</button>
-            </a>
-            <button ng-click="saveUser(userForm)" ng-disabled="userForm.$invalid" class="btn btn-success btn-primary pull-right">Salvar</button>
-        </div>
-    </div>
-
-</form>
+<div class="divpaddingTop"></div><form name="userForm" novalidate><div class="panel panel-default"><div class="panel-heading"><h3 class="panel-title">Cadastro de Usuário</h3></div><div class="panel-body"><div class="panel panel-warning"><div class="panel-heading"><h3 class="panel-title">Dados pessoais</h3></div><div class="panel-body"><input type="hidden" name="id" ng-model="user.id"><div class="form-group"><label name="name" class="nome required control-label">Nome</label><input type="text" name="name" ng-model="user.name" class="string required form-control" placeholder="Nome completo" maxlength="300" required><p ng-show="userForm.name.$invalid && !userForm.name.$pristine" class="help-block">O nome é obrigatório.</p></div><div class="form-group"><label name="email" class="email required control-label">Email</label><input type="text" name="email" ng-model="user.email" class="string email required form-control" placeholder="correio eletrônico/email" maxlength="300" required><p ng-show="userForm.email.$invalid && !userForm.email.$pristine" class="help-block">O e-mail é obrigatório.</p></div><div class="form-group"><label name="phone" class="email required control-label">Telefone</label><input type="text" name="phone" ng-model="user.phone" class="string required form-control" placeholder="Telefone" maxlength="255" required><p ng-show="userForm.phone.$invalid && !userForm.phone.$pristine" class="help-block">O telefone é obrigatório.</p></div><div class="form-group"><label name="sex" class="email required control-label">Sexo</label><select name="sex" ng-model="user.sex" class="string required form-control" required><option value="">-</option><option value="Masculino">Masculino</option><option value="Feminino">Feminino</option></select><p ng-show="userForm.sex.$invalid && !userForm.sex.$pristine" class="help-block">O sexo é obrigatório.</p></div></div></div></div></div><div class="footer"><div class="row"><a href="#home"><button class="btn btn-default btn-primary pull-left">Voltar</button> </a><button ng-click="saveUser(userForm)" ng-disabled="userForm.$invalid" class="btn btn-success btn-primary pull-right">Salvar</button></div></div></form>
