--- test-ui/src/main/resources/static/gulp/gulpfile.js
@@ -1,8 +1,0 @@
-const gulp = require('gulp');
-const requireDir = require('require-dir');
-
-requireDir('./tasks');
-
-gulp.task('default', ['browser-sync'], function () {
-    
-});
